Fix Module Dependency chain for US city zipcodes module so that it is properly executed
