+++ application/main.py
@@ -59,7 +59,6 @@ def execute(entities):
 if __name__ == '__main__':
     root_entities = [
         entity_key.census_us_metro_area,
-        entity_key.census_us_city,
         entity_key.census_us_city_zipcode,
         entity_key.snap_nyc_data,
         entity_key.mvc_nyc_crashes,
@@ -78,7 +77,7 @@ if __name__ == '__main__':
         entity_key.police_fatal_shootings,
         entity_key.apha_racism_declarations,
         # entity_key.ucla_covid_behind_bars,
-        entity_key.osha_closed_complaints,
+        # entity_key.osha_closed_complaints,
         # entity_key.twitter_account,
         # entity_key.twitter_tweets,
         # entity_key.twitter_tweets_terms,
