+++ census/employment.py
@@ -1,5 +1,6 @@
 from common.constants import entity_key
 from common.http import send_request
+from common.utils import execute_threads
 from entity.abstract import ResourceEntity
 
 import threading
@@ -10,16 +11,6 @@ THREAD_POOL_LIMIT = 25
 
 def build_url_query(census_tracts):
     return ':'.join(['1400000US' + tract['fips'] for tract in census_tracts])
-
-
-def execute_threads(threads):
-    for thread in threads:
-        thread.start()
-
-    for thread in threads:
-        thread.join()
-
-    threads.clear()
 
 
 class EmploymentStatus(ResourceEntity):
@@ -62,30 +53,6 @@ class EmploymentStatus(ResourceEntity):
 
         # Execute the threads that may be remaining
         execute_threads(threads)
-
-        # field_index = 349
-        # employed_index = 685
-        # unemployed_index = 776
-        # print(f'Calling endpoint {EmploymentStatus.API_URL}{build_url_query(batch_records)}')
-        # response = requests.request('GET', f'{EmploymentStatus.API_URL}{build_url_query(batch_records)}')
-        # response_content = response.content.decode('utf-8')
-        # try:
-        #     records = json.loads(response_content)
-        #     batch_record_count = len(batch_records)
-        #     batch_index = 0
-        #     while batch_index < batch_record_count:
-        #         data = records['response']['data']
-        #         if batch_index + 1 < len(data):
-        #             self.records.append({
-        #                 'census_tract_id': batch_records[batch_index]['id'],
-        #                 'employed': int(data[batch_index + 1][employed_index]),
-        #                 'unemployed': int(data[batch_index + 1][unemployed_index])
-        #             })
-        #         batch_index += 1
-        # except JSONDecodeError:
-        #     print('Cannot decode response content')
-        #     print(response_content)
-        #     census_tract_index += 1
 
     def async_fetch(self, batch_records):
         employed_index = 685
